refactor(cfd-3d): Decompose cascade/mod.rs into types and solver leaf modules
Part of ATLAS-ARCH-007.
